Unprotect imagery search page
